+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
-import Hero from "./Hero";
-import Navbar from "./navbar";
-import Rooms from "./rooms";
+import Hero from "./components/Hero";
+import Navbar from "./components/navbar";
+import Rooms from "./components/rooms";
 
 function App() {
   return (
