+++ src/components/HeroText.js
